+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "cyezoi" extension will be documented in this file.
 
 ## v0.0.19
 
-- **Fix attend problem**
+- **Automatically refresh the record tree view when the record is updated**
+- Fix attend problem
 
 Want to see the previous log? [Click here](https://github.com/CYEZOI/cyezoi-helper/commits/main/CHANGELOG.md)